Move CRF to preprocessing
