+++ core_analysis/architecture.py
@@ -6,12 +6,6 @@ import numpy as np
 import tensorflow as tf
 from keras import callbacks
 from keras import backend as K
-import pydensecrf.densecrf as dcrf
-from pydensecrf.utils import (
-    create_pairwise_gaussian,
-    create_pairwise_bilateral,
-    unary_from_softmax,
-)
 import segmentation_models as sm
 
 from core_analysis.postprocess import predict_tiles
@@ -81,56 +75,6 @@ class Model:
         pred_tile.predict(batches_num=1500, coords_channels=False)
         results = pred_tile.merge()
         return results
-
-
-def dense_crf(image, final_probabilities, gw=11, bw=3, n_iterations=5):
-    """
-    gw - pairwise gaussian window size: enforces more spatially consistent segmentations.
-    bw - pairwise bilateral window size: uses local color features to refine predictions.
-    """
-
-    ny = image.shape[0]
-    nx = image.shape[1]
-    n_classes = final_probabilities.shape[-1]
-    softmax = final_probabilities.squeeze()
-    softmax = softmax.transpose((2, 0, 1))
-
-    # The input should be the negative of the logarithm of probability values.
-    # Look up the definition of the unary_from_softmax for more information.
-    unary = unary_from_softmax(softmax, scale=None, clip=1e-5)
-
-    # The inputs should be C-continious -- we are using Cython wrapper.
-    unary = np.ascontiguousarray(unary)
-
-    d = dcrf.DenseCRF(ny * nx, n_classes)
-
-    d.setUnaryEnergy(unary)
-
-    # This potential penalizes small pieces of segmentation that are
-    # spatially isolated -- enforce more spatially consistent segmentations.
-    feats = create_pairwise_gaussian(sdims=(gw, gw), shape=(ny, nx))
-
-    d.addPairwiseEnergy(
-        feats, compat=3, kernel=dcrf.DIAG_KERNEL, normalization=dcrf.NORMALIZE_SYMMETRIC
-    )
-
-    # This creates the color-dependent features --
-    # because the segmentation that we get from CNN are too coarse
-    # and we can use local color features to refine them.
-    feats = create_pairwise_bilateral(
-        sdims=(bw, bw), schan=(7, 7, 7), img=image, chdim=2
-    )
-
-    d.addPairwiseEnergy(
-        feats, compat=3, kernel=dcrf.DIAG_KERNEL, normalization=dcrf.NORMALIZE_SYMMETRIC
-    )
-
-    Q = d.inference(n_iterations)
-    probs = np.array(Q, dtype=np.float32).reshape((n_classes, ny, nx))
-    probs = np.around(probs, 4)
-    # res = np.argmax(Q, axis=0).reshape((ny, nx))
-
-    return probs.swapaxes(1, 0).swapaxes(1, 2)
 
 
 class masked_loss:
